fix:[#20] uuid -> unique 컬럼 변경및 설문 중복 참여 로직 구현

+++ tellme/src/main/java/com/tellme/tellme/domain/survey/entity/SurveyShortUrl.java
@@ -15,11 +15,11 @@ public class SurveyShortUrl extends BaseEntity {
     @Column(name = "survey_id")
     private int surveyId;
     @Column(name = "user_id")
-    private long userId;
+    private int userId;
     private String url;
 
     @Builder
-    public SurveyShortUrl(int surveyId, long userId, String url) {
+    public SurveyShortUrl(int surveyId, int userId, String url) {
         this.surveyId = surveyId;
         this.userId = userId;
         this.url = url;
